eco-modelo-yolov5n/requirements.txt
grpcio==1.64.0
grpcio-tools==1.64.0
ultralytics>=8.0.0
opencv-python
numpy
torch
torchvision
yolov5